EC-19 style table

+++ routes/web.php
@@ -31,13 +31,16 @@ Route::group(['prefix' => 'user'], function () {
     Route::put('/edit/{slug}', [UserController::class, 'edit'])->name('edituser');
 });
 
-Route::group(['prefix' => 'order'], function () {
-    Route::get('/all', [OrderController::class, 'index'])->name('order.index');
-    Route::get('/create', [OrderController::class, 'create'])->name('order.create');
-    Route::post('/save', [OrderController::class, 'store'])->name('order.store');
-    Route::get('/show/{order}', [OrderController::class, 'show'])->name('order.show');
-    Route::get('/edit/{order}', [OrderController::class, 'edit'])->name('order.edit');
-    Route::put('/update/{order}', [OrderController::class, 'update'])->name('order.update');
-    Route::delete('/destroy/{order}', [OrderController::class, 'destroy'])->name('order.destroy');
+Route::group(['prefix' => 'admin/order'], function () {
+    Route::get('/all', [OrderController::class, 'index'])->name('admin.order.index');
+    Route::get('/create', [OrderController::class, 'create'])->name('admin.order.create');
+    Route::post('/save', [OrderController::class, 'store'])->name('admin.order.store');
+    Route::get('/show/{order}', [OrderController::class, 'show'])->name('admin.order.show');
+    Route::get('/edit/{order}', [OrderController::class, 'edit'])->name('admin.order.edit');
+    Route::put('/update/{order}', [OrderController::class, 'update'])->name('admin.order.update');
+    Route::delete('/destroy/{order}', [OrderController::class, 'destroy'])->name('admin.order.destroy');
 
 });
+
+
+
